The BMat8 type

This page contains the documentation of the type BMat8.

Contents

Function	Description

BMat8	Construct from Vector{Vector{T}} where {T<:Union{Bool,Int64}}

col_space_basis	Find a basis for the column space of a BMat8.

col_space_size	Returns the size of the column space of a BMat8.

degree	Returns the degree of a BMat8.

number_of_cols	Returns the number of non-zero columns in a BMat8.

number_of_rows	Returns the number of non-zero rows in a BMat8.

is_regular_element	Check whether a BMat8 is regular.

minimum_dim	Returns the minimum dimension of a BMat8.

one	Returns the identity BMat8 of a given dimension.

one	Returns the identity BMat8 of a given dimension.

random	Returns a random BMat8 of a given dimension.

random	Returns a random BMat8 of a given dimension.

row_space_basis	Find a basis for the row space of a BMat8.

row_space_size	Returns the size of the row space of a BMat8.

rows	Returns the rows of a BMat8.

swap!	Swaps two BMat8 objects.

to_int	Returns the integer representation of a BMat8.

Full API

Semigroups.BMat8 — Type

Fast boolean matrices of dimension up to 8 x 8.

This class represents 8 x 8 matrices over the boolean semiring. The functions for these small matrices over the boolean semiring are more optimised than the generic functions for boolean matrices. Note that all BMat8 are represented internally as an 8 x 8 matrix; any entries not defined by the user are taken to be 0. This does not affect the results of any calculations.

Example

julia> using Semigroups

julia> x = BMat8([[0, 1], [1, 0]])
BMat8([[0, 1],
       [1, 0]])

julia> x[1, 1] = 1
1

julia> x
BMat8([[1, 1],
       [1, 0]])

julia> x[1, 2]
true

julia> x[2, 2]
false

julia> x * x
BMat8([[1, 1],
       [1, 1]])

julia> x < x * x
true

julia> x *= x
BMat8([[1, 1],
       [1, 1]])

julia> x
BMat8([[1, 1],
       [1, 1]])

julia> to_int(x)
0xc0c0000000000000

julia> string(to_int(x), base=2)
"1100000011000000000000000000000000000000000000000000000000000000"

julia> x == BMat8([[1, 1, 0], [1, 1, 0], [0, 0, 0]]) # All BMat8's are really 8x8!
true

julia> y = BMat8([[1, 0, 1], [0, 1, 0], [0, 0, 0]])
BMat8([[1, 0, 1],
       [0, 1, 0],
       [0, 0, 0]])

julia> y[1]  # The first row
8-element Vector{Bool}:
 1
 0
 1
 0
 0
 0
 0
 0

julia> x + y
BMat8([[1, 1, 1],
       [1, 1, 0],
       [0, 0, 0]])

julia> x += y
BMat8([[1, 1, 1],
       [1, 1, 0],
       [0, 0, 0]])

julia> 1 * x == x
true

julia> x * 0
BMat8(0)

julia> Dict(BMat8(0) => 1)
Dict{Semigroups.LibSemigroups.BMat8Allocated, Int64} with 1 entry:
  BMat8(0) => 1

BMat8 objects can be used with the following algorithms in Semigroups.jl

Action

FroidurePin

Konieczny

Semigroups.BMat8 — Method

BMat8(rows::Vector{Vector{T}})::BMat8 where {T<:Union{Bool,Int64}} -> BMat8

Construct from Vector of rows.

This constructor initializes a matrix where the rows of the matrix are the lists in rows.

Arguments

rows::Vector{Vector{T}} where {T<:Union{Bool,Int64}}: the rows of the matrix.

Throws

MethodError: if rows has 0 rows.

LibsemigroupsError: if rows has more than 8 rows.

LibsemigroupsError: if the items in rows are not all of the same length.

Complexity

Constant.

Semigroups.col_space_basis — Method

 col_space_basis(x::BMat8) -> BMat8

Find a basis for the column space of a BMat8.

This function returns a BMat8 whose non-zero columns form a basis for the column space of x.

Arguments

x::BMat8: the matrix.

Complexity

Constant.

Example

julia> using Semigroups

julia> x = BMat8([[1, 0, 1], [0, 1, 0], [0, 0, 0]])
BMat8([[1, 0, 1],
       [0, 1, 0],
       [0, 0, 0]])

julia> col_space_basis(x)
BMat8([[1, 0],
       [0, 1]])

Semigroups.col_space_size — Method

 col_space_size(x::BMat8) -> Int64

Returns the size of the column space of a BMat8.

Arguments

x::BMat8: the matrix.

Complexity

$O(n)$ where $n$ is the return value of this function.

See Also

row_space_size.

Example

julia> using Semigroups

julia> x = BMat8([[0, 1], [1, 0]])
BMat8([[0, 1],
       [1, 0]])

julia> col_space_size(x)
4

Semigroups.degree — Method

degree(x::BMat8) -> Int64

Returns the degree of a BMat8.

This function returns the degree of x which always returns 8.

Arguments

x::BMat8: the matrix.

Semigroups.is_regular_element — Method

is_regular_element(x::BMat8) -> Bool

Check whether x is a regular element of the full boolean matrix monoid of appropriate dimension.

This function returns true if there exists a boolean matrix y such that x * y * x = x where x, and false otherwise.

Arguments

x::BMat8: the matrix.

Complexity

Constant.

Example

julia> using Semigroups

julia> x = BMat8([[0, 1], [1, 0]])
BMat8([[0, 1],
       [1, 0]])

julia> is_regular_element(x)
true

julia> sum(1 for x in 0:100000 if is_regular_element(BMat8(x)))
97997

Semigroups.minimum_dim — Method

minimum_dim(x::BMat8) -> Int64:

Returns the minimum dimension of a BMat8.

This function returns the maximal n such that row n or column n in the boolean matrix x contains at least one occurrence of 1. Equivalent to the maximum of number_of_rows and number_of_cols.

Arguments

x::BMat8: the matrix.

Complexity

Constant.

Example

julia> using Semigroups   

julia> x = BMat8([[0, 1], [1, 0]])
BMat8([[0, 1],
       [1, 0]])

julia> minimum_dim(x)
2

Semigroups.number_of_cols — Method

number_of_cols(x::BMat8) -> Int64

Returns the number of non-zero columns in a BMat8.

BMat8 objects do not know their "dimension" - in effect they are all of dimension 8. However, this function can be used to obtain the number of non-zero columns of a BMat8.

Arguments

x::BMat8: the matrix.

Complexity

Constant.

See also

number_of_rows(::BMat8) and minimum_dim(::BMat8).

Example

julia> using Semigroups   

julia> x = BMat8([[1, 0, 1], [0, 1, 0], [0, 0, 0]])
BMat8([[1, 0, 1],
       [0, 1, 0],
       [0, 0, 0]])

julia> number_of_cols(x)
3

Semigroups.number_of_rows — Method

number_of_rows(x::BMat8) -> Int64

Returns the number of non-zero rows in a BMat8.

BMat8 objects do not know their "dimension" - in effect they are all of dimension 8. However, this function can be used to obtain the number of non-zero rows of a BMat8.

Arguments

x::BMat8: the matrix.

Complexity

Constant.

See also

number_of_cols(::BMat8) and minimum_dim(::BMat8).

Example

julia> using Semigroups   

julia> x = BMat8([[1, 0, 1], [0, 1, 0], [0, 0, 0]])
BMat8([[1, 0, 1],
       [0, 1, 0],
       [0, 0, 0]])

julia> number_of_rows(x)
2

Base.one — Method

one(::Type{BMat8}, dim::Int64) -> BMat8

Returns the identity BMat8 of a given dimension.

This function returns the BMat8 with the first dim entries in the main diagonal equal to 1 and every other value equal to 0.

Arguments

dim::Int64: the dimension.

Throws

LibsemigroupsError: if dim is not in the interval $[0, 8]$.

Complexity

Constant.

Example

julia> using Semigroups

julia> one(BMat8, 4)
BMat8([[1, 0, 0, 0],
       [0, 1, 0, 0],
       [0, 0, 1, 0],
       [0, 0, 0, 1]])

Base.one — Method

one(sample::BMat8, dim::Int64) -> BMat8

Returns the identity BMat8 of a given dimension.

This function returns the BMat8 with the first dim entries in the main diagonal equal to 1 and every other value equal to 0.

Arguments

sample::BMat8: a matrix.

dim::Int64: the dimension.

Throws

LibsemigroupsError: if dim is not in the interval $[0, 8]$.

Complexity

Constant.

Example

julia> using Semigroups

julia> x = BMat8([[1, 0, 1], [0, 1, 0], [0, 0, 0]])
BMat8([[1, 0, 1],
       [0, 1, 0],
       [0, 0, 0]])

julia> one(x, 4)
BMat8([[1, 0, 0, 0],
       [0, 1, 0, 0],
       [0, 0, 1, 0],
       [0, 0, 0, 1]])

Semigroups.random — Method

random(::Type{BMat8}, dim::Int64) -> BMat8

Construct a random BMat8 of dimension at most dim.

This function returns a BMat8 chosen at random, where only the top-left dim by dim entries can be non-zero.

Arguments

dim::Int64: the dimension.

Throws

LibsemigroupsError: if dim is not in the interval $[1, 8]$.

Semigroups.random — Method

random(sample::BMat8, dim::Int64) -> BMat8

Construct a random BMat8 of dimension at most dim.

This function returns a BMat8 chosen at random, where only the top-left dim by dim entries can be non-zero.

Arguments

sample::BMat8: a matrix.

dim::Int64: the dimension.

Throws

LibsemigroupsError: if dim is not in the interval $[1, 8]$.

Semigroups.row_space_basis — Method

 row_space_basis(x::BMat8) -> BMat8

Find a basis for the row space of a BMat8.

This function returns a BMat8 whose non-zero rows form a basis for the row space of x.

Arguments

x::BMat8: the matrix.

Complexity

Constant.

Example

julia> using Semigroups

julia> x = BMat8([[1, 0, 1], [0, 1, 0], [0, 0, 0]])
BMat8([[1, 0, 1],
       [0, 1, 0],
       [0, 0, 0]])

julia> row_space_basis(x)
BMat8([[1, 0, 1],
       [0, 1, 0],
       [0, 0, 0]])

Semigroups.row_space_size — Method

 row_space_size(x::BMat8) -> Int64

Returns the size of the row space of a BMat8.

Arguments

x::BMat8: the matrix.

Complexity

$O(n)$ where $n$ is the return value of this function.

See Also

col_space_size.

Example

julia> using Semigroups

julia> x = BMat8([[0, 1], [1, 0]])
BMat8([[0, 1],
       [1, 0]])

julia> row_space_size(x)
4

Semigroups.rows — Method

rows(x::BMat8) -> Vector{Vector{Bool}}

Returns the rows of a BMat8.

This function returns the rows of x. The returned Vector always has length 8, even if x was constructed with fewer rows.

Arguments

x::BMat8: the matrix.

Complexity

Constant.

Example

julia> using Semigroups

julia> x = BMat8([[0, 1], [1, 0]])
BMat8([[0, 1],
       [1, 0]])

julia> rows(x)
8-element Vector{Vector{Bool}}:
 [0, 1, 0, 0, 0, 0, 0, 0]
 [1, 0, 0, 0, 0, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 0]
 [0, 0, 0, 0, 0, 0, 0, 0]

Semigroups.swap! — Method

swap!(x::BMat8, y::BMat8) -> nothing

Swaps two BMat8 objects.

This function swaps the values of x and y.

Arguments

x::BMat8: the first matrix to swap.

y::BMat8: the second matrix to swap.

Complexity

Constant.

Example

julia> using Semigroups

julia> x = BMat8([[0, 1], [1, 0]])
BMat8([[0, 1],
       [1, 0]])

julia> y = BMat8([[1, 1], [0, 0]])
BMat8([[1, 1],
       [0, 0]])

julia> swap!(x, y)

julia> x
BMat8([[1, 1],
       [0, 0]])

julia> y
BMat8([[0, 1],
       [1, 0]])

Semigroups.to_int — Method

to_int(x::BMat8) -> UInt64

Returns the integer representation of a BMat8.

This function returns a non-negative integer obtained by interpreting an 8 x 8 BMat8 as a sequence of 64 bits (reading rows left to right, from top to bottom) and then realising this sequence as an integer.

Arguments

x::BMat8: the matrix.

Complexity

Constant.

Example

julia> using Semigroups

julia> x = BMat8([[0, 1], [1, 0]])
BMat8([[0, 1],
       [1, 0]])

julia> to_int(x)
0x4080000000000000

julia> string(to_int(x), base=2)
"100000010000000000000000000000000000000000000000000000000000000"