The WordRange type

This page contains the documentation of the type WordRange, an iterable range over words in an alphabet in a user-selected Order.

Semigroups.WordRange — Type

WordRange

An iterable range over words in an alphabet, in a user-selected order.

Configure with set_alphabet_size!, set_min!, set_max!, set_first!, set_last!, set_order!, set_upper_bound!, then iterate with Base.get(::WordRange) / next! / at_end or the standard Julia iteration protocol (for w in r, collect(r), etc.).

A default-constructed WordRange is empty, with alphabet size 0, order ORDER_SHORTLEX, and empty first/last bookends.

Example

julia> using Semigroups

julia> r = WordRange();

julia> set_alphabet_size!(r, 2);

julia> set_min!(r, 1);

julia> set_max!(r, 3);

julia> collect(r)
6-element Vector{Vector{Int64}}:
 [1]
 [2]
 [1, 1]
 [1, 2]
 [2, 1]
 [2, 2]

Contents

Function	Description

alphabet_size	The current number of letters in the alphabet.

set_alphabet_size!	Set the number of letters in the alphabet.

at_end	Check if the range is exhausted.

count	The actual size of the range.

first_word	The current first word (lower bookend) of the range.

set_first!	Set the first word (lower bookend) of the range.

get	Get the current word from the range.

init!	Re-initialize an existing WordRange object.

last_word	The current one-past-the-last word (upper bookend) of the range.

set_last!	Set the one-past-the-last word (upper bookend) of the range.

set_max!	Set one past the maximum word length.

set_min!	Set the minimum word length.

next!	Advance the range to the next word.

order	The current order of the words in the range.

set_order!	Set the order of the words in the range.

size_hint	The possible size of the range.

upper_bound	The current upper bound on the length of a word in the range.

set_upper_bound!	Set an upper bound on the length of a word in the range.

valid	Whether settings have been changed since the last next! / get call.

Free functions

Function	Description

number_of_words	Count the words over an alphabet of size n with length in [min, max).

random_word	Return a uniformly random word of the given length over an alphabet of a given size.

Full API

Semigroups.alphabet_size — Method

alphabet_size(r::WordRange) -> Int

Return the current number of letters in the alphabet of r.

Semigroups.set_alphabet_size! — Method

set_alphabet_size!(r::WordRange, n::Integer) -> WordRange

Set the number of letters in the alphabet of r to n.

Arguments

n::Integer: the number of letters.

Semigroups.at_end — Method

at_end(r::WordRange) -> Bool

Check if the range has been exhausted.

Returns true if a WordRange object is exhausted, and false otherwise.

Base.count — Method

count(r::WordRange) -> Int

The actual size of the range.

Returns the number of words in a WordRange object. If the order is ORDER_SHORTLEX then size_hint is used. If the order is not shortlex then a copy of the range may have to be looped over in order to compute the return value.

Semigroups.first_word — Method

first_word(r::WordRange) -> Vector{Int}

Return the current first word (lower bookend) of r.

Returned as a Vector{Int} of 1-based letter indices. See set_first! to configure.

Semigroups.set_first! — Method

set_first!(r::WordRange, w::AbstractVector{<:Integer}) -> WordRange

Set the first word (lower bookend) of r to w.

Performs no checks on its arguments. If w contains letters greater than alphabet_size, then r will be empty. Similarly, if the first word is greater than the last word with respect to the current order, r will be empty.

Arguments

w::AbstractVector{<:Integer}: the first word (1-based letter indices).

Base.get — Method

Base.get(r::WordRange) -> Vector{Int}

Get the current word from the range.

Returns the current word in a WordRange object as a Vector{Int} of 1-based letter indices.

Semigroups.init! — Method

init!(r::WordRange) -> WordRange

Re-initialize an existing WordRange object.

Puts r back into the same state as if it had been newly default-constructed.

Semigroups.last_word — Method

last_word(r::WordRange) -> Vector{Int}

Return the current one-past-the-last word (upper bookend) of r.

Returned as a Vector{Int} of 1-based letter indices. See set_last! to configure.

Semigroups.set_last! — Method

set_last!(r::WordRange, w::AbstractVector{<:Integer}) -> WordRange

Set the one-past-the-last word (upper bookend) of r to w.

Performs no checks on its arguments. If w contains letters greater than alphabet_size, then r will be empty.

Arguments

w::AbstractVector{<:Integer}: one past the last word (1-based letter indices).

Semigroups.set_max! — Method

set_max!(r::WordRange, n::Integer) -> WordRange

Set one past the maximum word length of r to n.

Sets the one-past-the-last word of r to the word consisting of n copies of letter 1. So, after set_max!(r, n), r contains words of length strictly less than n.

Arguments

n::Integer: one greater than the maximum length.

Semigroups.set_min! — Method

set_min!(r::WordRange, n::Integer) -> WordRange

Set the minimum word length of r to n.

Sets the first word of r to the word consisting of n copies of letter 1.

Arguments

n::Integer: the minimum length.

Semigroups.next! — Method

next!(r::WordRange) -> WordRange

Advance the range to the next word (if any).

See also

at_end

AbstractAlgebra.order — Method

order(r::WordRange) -> Order

Return the current Order of the words in r.

Semigroups.set_order! — Method

set_order!(r::WordRange, o::Order) -> WordRange

Set the order of the words in r to o.

Arguments

o::Order: the order; must be ORDER_SHORTLEX or ORDER_LEX.

Throws

LibsemigroupsError: if o is not ORDER_SHORTLEX or ORDER_LEX.

Semigroups.size_hint — Method

size_hint(r::WordRange) -> Int

The possible size of the range.

Returns the number of words in a WordRange object if its order is ORDER_SHORTLEX. If the order is not shortlex, then the return value of this function is meaningless.

Semigroups.upper_bound — Method

upper_bound(r::WordRange) -> Int

Return the current upper bound on the length of a word in r.

This setting is only used when the order of r is ORDER_LEX.

Semigroups.set_upper_bound! — Method

set_upper_bound!(r::WordRange, n::Integer) -> WordRange

Set an upper bound for the length of a word in r.

This setting is only used when the order of r is ORDER_LEX.

Arguments

n::Integer: the upper bound.

Semigroups.valid — Method

valid(r::WordRange) -> Bool

Return whether the settings have not changed since the last call to next! or Base.get(::WordRange).

Other than by calling next!, the value returned by Base.get(::WordRange) may be altered by a call to any of: set_order!, set_alphabet_size!, set_min!, set_max!, set_first!, set_last!, set_upper_bound!.

Returns true if none of these settings have changed since the last next! or Base.get(::WordRange) call, and false otherwise.

Semigroups.number_of_words — Method

number_of_words(n::Integer, min::Integer, max::Integer) -> Int

Return the number of words over an alphabet of n letters with length in the range [min, max), i.e. $\sum_{i=\min}^{\max-1} n^i$.

Arguments

n::Integer: the number of letters in the alphabet.

min::Integer: the minimum word length.

max::Integer: one greater than the maximum word length.

Warning

If the resulting count exceeds 2^64 - 1 the return value will not be correct.

Semigroups.random_word — Method

random_word(length::Integer, nr_letters::Integer) -> Vector{Int}

Return a uniformly random word of the given length over an alphabet of nr_letters letters, with 1-based letter indices.

Arguments

length::Integer: the length of the word.

nr_letters::Integer: the size of the alphabet.

Throws

LibsemigroupsError: if nr_letters is 0.