The Runner class

Semigroups.Runner — Type

Runner

Abstract class for derived classes that run an algorithm.

Many of the classes in Semigroups.jl implementing the algorithms that are the reason for the existence of this library, are derived from Runner. The Runner class exists to collect various common tasks required by such a derived class with a possibly long running run!.

These common tasks include:

running for a given amount of time (run_for!)

running until a nullary predicate is true (run_until!)

checking the status of the algorithm: has it started? finished? been killed by another thread (dead)? has it timed out (timed_out)? has it stopped for any reason?

permit the function run! to be killed from another thread (kill!).

Contents

Function	Description

current_state	Return the current state.

dead	Check if the runner is dead.

finished	Check if run has been run to completion or not.

init!	Initialize an existing Runner object.

kill!	Stop run from running (thread-safe).

report_why_we_stopped	Report why run stopped.

run!	Run until finished.

run_for!	Run for a specified amount of time.

run_until!	Run until a nullary predicate returns true or finished.

running	Check if currently running.

running_for	Check if the runner is currently running for a particular length of time.

running_until	Check if the runner is currently running until a nullary predicate returns true.

started	Check if run has been called at least once before.

stopped	Check if the runner is stopped.

stopped_by_predicate	Check if the runner was stopped, or should stop, because of the argument last passed to run_until!.

success	Check if run has been run to completion successfully.

timed_out	Check if the amount of time passed to run_for has elapsed.

Full API

Semigroups.current_state — Method

current_state(r::Runner) -> RunnerState

Return the current state.

Returns the current state of the Runner as given by state.

Semigroups.dead — Method

dead(r::Runner) -> Bool

Check if the runner is dead.

This function can be used to check if we should terminate run! because it has been killed by another thread.

See also

kill!

Semigroups.finished — Method

finished(r::Runner) -> Bool

Check if run!(::Runner) has been run to completion or not.

Returns true if run! has been run to completion; and false if not.

See also

Semigroups.init! — Method

init!(r::Runner) -> Runner

Initialize an existing Runner object.

This function puts a Runner object back into the same state as if it had been newly default constructed.

Note

This function is not thread-safe.

See also

Runner

Semigroups.kill! — Method

kill!(r::Runner)

Stop run! from running (thread-safe).

This function can be used to terminate run! from another thread. After kill! has been called the Runner may no longer be in a valid state, but will return true from dead .

See also

finished

Semigroups.report_why_we_stopped — Method

report_why_we_stopped(r::Runner)

Report why run! stopped.

Reports whether run! was stopped because it is finished, timed_out, or dead.

Semigroups.run! — Method

run!(r::Runner)

Run until finished.

Run the main algorithm implemented by a derived class of Runner.

Semigroups.run_for! — Method

run_for!(r::Runner, t::TimePeriod)

Run for a specified amount of time.

For this to work it is necessary to periodically check if timed_out returns true, and to stop if it is, in the run! method of any subtype of Runner.

Arguments

t::TimePeriod: the duration to run for.

Examples

run_for!(r, Second(1))
run_for!(r, Millisecond(500))

Semigroups.run_until! — Method

run_until!(r::Runner, f::Function)

Run until a nullary predicate returns true or finished.

This function runs the algorithm until the nullary predicate f returns true or the algorithm finished.

Arguments

f::Function: a function with 0 arguments returning true or false

Supports do-block syntax:

run_until!(r) do
    some_condition(r)
end

Semigroups.running — Method

running(r::Runner) -> Bool

Check if currently running.

Returns true if run! is in the process of running and false it is not.

See also

finished

Semigroups.running_for — Method

running_for(r::Runner) -> Bool

Check if the runner is currently running for a particular length of time.

If the Runner is currently running because its member function run_for! has been invoked, then this function returns true. Otherwise, false is returned.

Semigroups.running_until — Method

running_until(r::Runner) -> Bool

Check if the runner is currently running until a nullary predicate returns true.

If the Runner is currently running because its member function run_until! has been invoked, then this function returns true. Otherwise, false is returned.

Semigroups.started — Method

started(r::Runner) -> Bool

Check if run! has been called at least once before.

Returns true if run! has started to run (it can be running or not).

See also

finished

Semigroups.stopped — Method

stopped(r::Runner) -> Bool

Check if the runner is stopped.

This function can be used to check whether or not run! has been stopped for whatever reason. In other words, it checks if timed_out, finished, or dead.

Semigroups.stopped_by_predicate — Method

stopped_by_predicate(r::Runner) -> Bool

Check if the runner was stopped, or should stop, because of the argument last passed to run_until!.

If r is running, then the nullary predicate is called and its return value is returned. If r is not running, then true is returned if and only if the last time r was running it was stopped by a call to the nullary predicate passed to run_until!.

Base.success — Method

Base.success(r::Runner) -> Bool

Check if run! has been run to completion successfully.

Returns true if run! has been run to completion and it was successful. The default implementation is to just call finished.

This extends Base.success (which checks process exit status) to work with libsemigroups Runner types.

Semigroups.timed_out — Method

timed_out(r::Runner) -> Bool

Check if the amount of time passed to run_for! has elapsed.

See also

run_for!(::Runner, ::TimePeriod)